hotfix

+++ src/renderlib/Framebuffer.cpp
@@ -6,7 +6,7 @@
 Framebuffer::Framebuffer(): width(100), height(100), fbStorage(100 * 100) { }
 Framebuffer::Framebuffer(int w, int h): width(w), height(h), fbStorage(w * h) { }
 
-int Framebuffer::floatToPngColor(float c) { return static_cast<int>(c * 255.0f); }
+float Framebuffer::floatToPngColor(float c) { return (c * 255.0f); }
 
 
 void Framebuffer::setPixelColor(int i, vec3 color) {
